+++ data/db-init.sql
@@ -1,0 +1,5 @@
+﻿CREATE DATABASE OmegaUnitTests;
+
+CREATE TABLE TestEntity (
+    
+)